Move handler to a shared package
