+++ pkg/clientcfg/clientcfghttp/handler.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package httpserver
+package clientcfghttp
 
 import (
 	"encoding/json"
